requirements.txt
python-telegram-bot>=21.10
python-dotenv>=1.0.1
prisma>=6.14.0 
prisma-client>=6.14.0